requirements.txt
requests
pytelegrambotapi
python-dotenv